Move app error handlers into error blueprint
