+++ response_operations_ui/views/errors.py
@@ -1,8 +1,8 @@
 import logging
 
-from flask import Blueprint, render_template
+from flask import Blueprint, flash, redirect, render_template, url_for
 from structlog import wrap_logger
-from response_operations_ui.exceptions.exceptions import UpdateContactDetailsException
+from response_operations_ui.exceptions.exceptions import ApiError, UpdateContactDetailsException
 
 
 logger = wrap_logger(logging.getLogger(__name__))
@@ -25,3 +25,28 @@ def update_details_exception(error=None):
                            form=error.form,
                            error_type=error_type,
                            respondent_details=error.respondent_details)
+
+
+@error_bp.app_errorhandler(ApiError)
+def api_error(error):
+    logger.error('Api failed to retrieve required data', message=error.message, url=error.url,
+                 status=str(error.status_code))
+    return redirect(url_for('error_bp.server_error_page'))
+
+
+@error_bp.app_errorhandler(401)
+def handle_authentication_error(error):
+    logger.warn('Authentication failed')
+    flash('Incorrect username or password', category='failed_authentication')
+    return redirect(url_for('sign_in_bp.sign_in'))
+
+
+@error_bp.app_errorhandler(Exception)
+def server_error(error):  # pylint: disable=unused-argument
+    logger.exception('Generic exception generated')
+    return redirect(url_for('error_bp.server_error_page'))
+
+
+@error_bp.app_errorhandler(500)
+def server_error_500(error):  # pylint: disable=unused-argument
+    return redirect(url_for('error_bp.server_error_page'))
